refactor(hindsight): unify decoding behind TransferLedger + Strategy
Netting, Relay rebalance anchoring, and fee accounting each built their
own transfer tallies from logs and native transfers; they are now
queries on one TransferLedger built once per transaction, documented
with the netting model and its assumptions. Each Strategy variant
decodes through a single DecodeContext seam, the LiFi bridge veto moved
into venue matching, and post-decode guards live in their own module,
so the block orchestrator carries no venue-specific logic.

Decoded output is unchanged: 25 blocks sampled across the run5 range
decode field-identical to the run5 records.
